Make PatchableRecordSet the only RecordSet
