+++ data/README.md
@@ -804,6 +804,31 @@ const store = new Store({
 Stores connected to a Cube `View` need no configuration here - the View installs a digest on
 them automatically, reading a stamp it maintains on every row it publishes.
 
+### Incremental Patching for Large Datasets
+
+By default, every Store change - a transaction, reload, or refilter - rebuilds the affected
+record collections in full, an O(total records) cost however small the change. For large stores
+receiving small, frequent changes, the experimental `maxPatchRatio` config expresses changes as
+small *patch layers* over a shared base collection instead, making transaction, filtering, and
+grid-sync costs scale with the size of the change:
+
+```typescript
+const store = new Store({
+    experimental: {maxPatchRatio: 0.1} // patches capped at 10% of total records
+});
+```
+
+A patch that grows past the configured fraction of total records is flattened away (clamped to
+`[0, 0.5]` - the default of `0` disables patching). The ratio is read live on each operation, so it
+may also be changed on an existing Store at any time, and can be applied app-wide via the
+`xhStoreExperimental` soft-config. Two behavioral notes:
+
+- Record order becomes stable-by-incumbency rather than source-order: existing records keep their
+  positions and additions append, including adds within partial reloads and records entering a
+  filter incrementally. Apply a grid sort where deterministic order matters.
+- Incremental refiltering assumes filter outcomes depend only on each record's own data. Call
+  `Store.refreshFilter()` if a filter's results can change for external reasons.
+
 ### Tuning Memory for Large Datasets
 
 For stores holding tens of thousands of records or more, two independent knobs reduce retained
