+++ data/impl/RecordSet.ts
@@ -7,13 +7,17 @@
 
 import equal from 'fast-deep-equal';
 import {logWarn, throwIf} from '@xh/hoist/utils/js';
-import {maxBy, isNil} from 'lodash';
+import {clamp, maxBy, isNil} from 'lodash';
 import {StoreRecord, StoreRecordId} from '../StoreRecord';
 import {Store} from '../Store';
 import {Filter} from '../filter/Filter';
 
 type StoreRecordMap = Map<StoreRecordId, StoreRecord>;
 type ChildRecordMap = Map<StoreRecordId, StoreRecord[]>;
+
+/** Patch-layer entry marking a base record as removed. */
+const TOMBSTONE = {} as StoreRecord;
+type PatchMap = Map<StoreRecordId, StoreRecord>;
 
 /**
  * Changes deriving one RecordSet from another, as computed by {@link RecordSet.diffFrom}.
@@ -28,9 +32,34 @@ export interface RecordSetDelta {
     remove: StoreRecord[];
 }
 
+// Default cap on patch size as a fraction of base - 0 disables the patch layer entirely, with
+// every change flattening into a fresh base. Overridable via
+// `StoreConfig.experimental.maxPatchRatio`.
+const DEFAULT_MAX_PATCH_RATIO = 0;
+
+// Upper clamp on the configurable ratio. Past ~half the store a patch stops being meaningfully
+// incremental, while worst-case tombstone retention grows as ratio/(1-ratio) of the live set.
+const MAX_PATCH_RATIO = 0.5;
+
 /**
  * Internal container for StoreRecord management within a Store.
  * Note this is an immutable object; its update and filtering APIs return new instances as required.
+ *
+ * A persistent (structurally shared) collection: each instance holds a `base` map - shared with
+ * related instances and never mutated - plus an optional small `patch` layer of changed entries
+ * (updated/added records, or TOMBSTONEs marking removals). Transactions merge patches at
+ * O(patch) cost rather than copying the full map, and two instances sharing a base can derive
+ * the exact delta between them at O(patch) via `diffFrom` - the basis for incremental filtering
+ * and grid transaction sync. Patches are capped at one layer deep: deriving from a patched set
+ * merges into a new single patch, and one invariant governs all paths - a patch never exceeds
+ * the configured fraction of its base (`experimental.maxPatchRatio`). Transactions
+ * crossing the cap flatten into a fresh base (amortized O(n)); reloads changing more than it
+ * simply adopt the incoming map as a new base.
+ *
+ * The ratio defaults to 0, disabling the patch layer entirely - every change flattens
+ * immediately, preserving classic full-copy behavior and source-order records. The ratio is read
+ * live on each operation, so it may be set on an existing Store at any time. See
+ * {@link StoreConfig.experimental} for the ordering implications of enabling patching.
  *
  * @internal
  */
@@ -39,19 +68,32 @@ export class RecordSet {
     count: number;
     rootCount: number;
 
-    private _recordMap: StoreRecordMap; // Map of all Records by id
+    /** Shared base map - never mutated once installed in a RecordSet. */
+    readonly base: StoreRecordMap;
+    /** Changed entries relative to `base` (TOMBSTONE = removed), or null for a flat set. */
+    readonly patch: PatchMap;
+
     private _childrenMap: ChildRecordMap; // children by parentId
     private _list: StoreRecord[]; // all records.
     private _rootList: StoreRecord[]; // root records.
     private _maxDepth: number;
+    private _filterSource: RecordSet = null; // source a filtered projection was built from
+    private _filter: Filter = null; // filter a projection was built with
 
     derivation: RecordSetDerivation = null;
 
-    constructor(store: Store, recordMap: StoreRecordMap = new Map()) {
+    constructor(
+        store: Store,
+        recordMap: StoreRecordMap = new Map(),
+        patch: PatchMap = null,
+        count: number = -1,
+        rootCount: number = -1
+    ) {
         this.store = store;
-        this._recordMap = recordMap;
-        this.count = recordMap.size;
-        this.rootCount = this.countRoots(recordMap);
+        this.base = recordMap;
+        this.patch = patch?.size ? patch : null;
+        this.count = count >= 0 ? count : recordMap.size;
+        this.rootCount = rootCount >= 0 ? rootCount : this.countRoots(recordMap);
     }
 
     get empty(): boolean {
@@ -59,7 +101,29 @@ export class RecordSet {
     }
 
     getById(id: StoreRecordId): StoreRecord {
-        return this._recordMap.get(id);
+        const {patch} = this;
+        if (patch) {
+            const v = patch.get(id);
+            if (v !== undefined) return v === TOMBSTONE ? undefined : v;
+        }
+        return this.base.get(id);
+    }
+
+    /** Iterate all records, reading through any patch layer without materializing a map. */
+    forEachRecord(fn: (rec: StoreRecord, id: StoreRecordId) => void) {
+        const {base, patch} = this;
+        if (!patch) {
+            base.forEach(fn);
+            return;
+        }
+        base.forEach((rec, id) => {
+            const v = patch.get(id);
+            if (v === undefined) fn(rec, id);
+            else if (v !== TOMBSTONE) fn(v, id);
+        });
+        patch.forEach((v, id) => {
+            if (v !== TOMBSTONE && !base.has(id)) fn(v, id);
+        });
     }
 
     getDescendantsById(id: StoreRecordId): StoreRecord[] {
@@ -82,16 +146,23 @@ export class RecordSet {
     isEqual(other: RecordSet): boolean {
         if (this.count !== other.count) return false;
 
-        for (const [id, rec] of this._recordMap) {
-            if (rec !== other.getById(id)) return false;
+        // Sharing a base, equality is answerable from the patches alone.
+        if (this.base === other.base) {
+            const delta = this.diffFrom(other);
+            return !delta.update.length && !delta.add.length && !delta.remove.length;
         }
 
-        return true;
+        let ret = true;
+        this.forEachRecord((rec, id) => {
+            if (rec !== other.getById(id)) ret = false;
+        });
+        return ret;
     }
 
     /**
-     * Changes that would derive this RecordSet from `prev` - the delta contract consumed by
-     * Grid to sync ag-Grid transactionally.
+     * Changes that would derive this RecordSet from `prev` - computed by comparing patch
+     * layers at O(patch) when both instances share a base map, by a full scan of both sets
+     * otherwise. An empty delta means identical content.
      */
     diffFrom(prev: RecordSet): RecordSetDelta {
         const update = [],
@@ -100,35 +171,78 @@ export class RecordSet {
 
         if (!prev) return {update, add: this.list, remove};
 
-        this._recordMap.forEach((rec, id) => {
-            const existing = prev.getById(id);
-            if (!existing) {
-                add.push(rec);
-            } else if (existing !== rec) {
-                update.push(rec);
-            }
-        });
+        const {base, patch} = this,
+            prevPatch = prev.patch;
 
-        if (this.count !== prev.count + add.length) {
-            prev._recordMap.forEach((rec, id) => {
-                if (!this.getById(id)) remove.push(rec);
+        // Scan compare
+        if (prev.base !== base) {
+            this.forEachRecord((rec, id) => {
+                const existing = prev.getById(id);
+                if (!existing) {
+                    add.push(rec);
+                } else if (existing !== rec) {
+                    update.push(rec);
+                }
+            });
+            if (this.count !== prev.count + add.length) {
+                prev.forEachRecord((rec, id) => {
+                    if (!this.getById(id)) remove.push(rec);
+                });
+            }
+            return {update, add, remove};
+        }
+
+        //... or patch compare
+        if (patch !== prevPatch) {
+            const prevEff = (id: StoreRecordId): StoreRecord => {
+                if (prevPatch) {
+                    const v = prevPatch.get(id);
+                    if (v !== undefined) return v === TOMBSTONE ? undefined : v;
+                }
+                return base.get(id);
+            };
+
+            patch?.forEach((v, id) => {
+                const curr = v === TOMBSTONE ? undefined : v,
+                    prevRec = prevEff(id);
+                if (curr === prevRec) return;
+                if (!curr) {
+                    if (prevRec) remove.push(prevRec);
+                } else if (prevRec) {
+                    update.push(curr);
+                } else {
+                    add.push(curr);
+                }
+            });
+
+            // Entries patched only in prev (possible when diffing non-derived siblings) - their
+            // effective value here comes straight from base.
+            prevPatch?.forEach((v, id) => {
+                if (patch?.has(id)) return;
+                const curr = base.get(id),
+                    prevRec = v === TOMBSTONE ? undefined : v;
+                if (curr === prevRec) return;
+                if (!curr) {
+                    if (prevRec) remove.push(prevRec);
+                } else if (prevRec) {
+                    update.push(curr);
+                } else {
+                    add.push(curr);
+                }
             });
         }
 
         return {update, add, remove};
     }
 
-    /**
-     * As `diffFrom`, but only provided if cheaply derivable (i.e. without a full scan).
-     * Always null for this implementation.
-     */
+    /** As `diffFrom`, but only when answerable at O(patch) - null on unrelated instances. */
     deltaFrom(prev: RecordSet): RecordSetDelta {
-        return null;
+        return this.hasDeltaFrom(prev) ? this.diffFrom(prev) : null;
     }
 
-    /** True if `deltaFrom` would answer. Always false for this implementation. */
+    /** True if `deltaFrom` would answer - i.e. this and `prev` share a base map. */
     hasDeltaFrom(prev: RecordSet): boolean {
-        return false;
+        return !!prev && prev.base === this.base;
     }
 
     //----------------------------------------------------------
@@ -137,12 +251,16 @@ export class RecordSet {
     // clients will never ask for list or tree representations.
     //----------------------------------------------------------
     get childrenMap(): ChildRecordMap {
-        if (!this._childrenMap) this._childrenMap = this.computeChildrenMap(this._recordMap);
+        if (!this._childrenMap) this._childrenMap = this.computeChildrenMap();
         return this._childrenMap;
     }
 
     get list(): StoreRecord[] {
-        if (!this._list) this._list = Array.from(this._recordMap.values());
+        if (!this._list) {
+            const list = [];
+            this.forEachRecord(rec => list.push(rec));
+            this._list = list;
+        }
         return this._list;
     }
 
@@ -170,10 +288,21 @@ export class RecordSet {
         return this.isEqual(target) ? target : this;
     }
 
-    // This implementation always filters in full - `prevFiltered` (the previous projection) is
-    // consulted only to count the changes for diagnostics.
-    withFilter(filter: Filter, prevFiltered: RecordSet): RecordSet {
+    // Caller can pass a `prevFiltered`, if it has one that has the identical filter applied.
+    withFilter(filter: Filter, prevFiltered?: RecordSet): RecordSet {
         if (!filter) return this;
+
+        const ret =
+            this.withFilterIncremental(filter, prevFiltered) ??
+            this.withFilterFull(filter, prevFiltered);
+        if (ret !== prevFiltered) {
+            ret._filterSource = this;
+            ret._filter = filter;
+        }
+        return ret;
+    }
+
+    private withFilterFull(filter: Filter, prevFiltered: RecordSet): RecordSet {
         const {store} = this,
             includeChildren = store.filterIncludesChildren,
             test = filter.getTestFn(store),
@@ -196,14 +325,14 @@ export class RecordSet {
                     }
                 });
             };
-            this._recordMap.forEach(rec => {
+            this.forEachRecord(rec => {
                 if (!isMarked(rec) && test(rec)) {
                     mark(rec);
                     markChildren(rec);
                 }
             });
         } else {
-            this._recordMap.forEach(rec => {
+            this.forEachRecord(rec => {
                 if (test(rec)) mark(rec);
             });
         }
@@ -243,31 +372,53 @@ export class RecordSet {
 
     withNewRecords(recordMap: StoreRecordMap): RecordSet {
         // Reuse existing StoreRecord object instances where possible.
-        // If reload changed nothing - preserve instance identity outright,
+        // If reload changed nothing - preserve instance identity outright.
         // Be sure to finalize any new records that are accepted.
-        let reused = 0,
-            adds = 0;
+        const changed: StoreRecord[] = []; // accepted new instances - updates and adds
+        let adds = 0,
+            rootCount = 0;
         recordMap.forEach((newRec, id) => {
             const currRec = this.getById(id);
             if (currRec && this.areRecordsEqual(currRec, newRec)) {
                 recordMap.set(id, currRec);
-                reused++;
+                if (currRec.parentId == null) rootCount++;
             } else {
-                if (!currRec) adds++;
                 newRec.finalize();
+                if (!currRec) adds++;
+                changed.push(newRec);
+                if (newRec.parentId == null) rootCount++;
             }
         });
 
-        const count = recordMap.size;
-        if (reused === count && reused === this.count) return this;
+        const count = recordMap.size,
+            removedCount = this.count - (count - adds);
+        if (!removedCount && !changed.length) return this;
 
-        const ret = new RecordSet(this.store, recordMap);
-        ret.derivation = {
-            type: 'full',
-            update: count - reused - adds,
-            add: adds,
-            remove: this.count - (count - adds)
-        };
+        // When reuse dominates, express the new set as a patch over the incumbent base -
+        // preserving base identity so consumers can derive the (small) reload delta. Otherwise
+        // the incoming map simply becomes a fresh base.
+        const {store, base, patch} = this,
+            ratio = this.patchRatio(),
+            changes = changed.length + removedCount,
+            counts = {update: changed.length - adds, add: adds, remove: removedCount};
+
+        if (changes <= ratio * count) {
+            const newPatch: PatchMap = patch ? new Map(patch) : new Map();
+            changed.forEach(rec => newPatch.set(rec.id, rec));
+            if (removedCount) {
+                this.forEachRecord((rec, id) => {
+                    if (!recordMap.has(id)) removeFromPatch(id, newPatch, base);
+                });
+            }
+            if (newPatch.size <= ratio * base.size) {
+                const ret = new RecordSet(store, base, newPatch, count, rootCount);
+                ret.derivation = {type: 'patched', ...counts};
+                return ret;
+            }
+        }
+
+        const ret = new RecordSet(store, recordMap, null, count, rootCount);
+        ret.derivation = {type: 'full', ...counts};
         return ret;
     }
 
@@ -276,41 +427,59 @@ export class RecordSet {
         add?: StoreRecord[];
         remove?: StoreRecordId[];
     }): RecordSet {
-        const {update, add, remove} = t;
+        const {update, add, remove} = t,
+            {base, patch} = this;
 
+        // Merge into a copy of the current patch - O(patch), not O(all records).
         // Be sure to finalize any new records that are accepted.
-        const newRecords = new Map(this._recordMap);
-
-        let missingRemoves = 0,
+        const newPatch: PatchMap = patch ? new Map(patch) : new Map();
+        let {count, rootCount} = this,
+            missingRemoves = 0,
             missingUpdates = 0;
+
+        // Effective record as of this point in the transaction.
+        const eff = (id: StoreRecordId): StoreRecord => {
+            const v = newPatch.get(id);
+            if (v !== undefined) return v === TOMBSTONE ? undefined : v;
+            return base.get(id);
+        };
 
         // 0) Removes - process first to allow delete-then-add-elsewhere-in-tree.
         if (remove) {
-            const allRemoves = new Set<StoreRecordId>();
+            const isTree = this.count !== this.rootCount,
+                allRemoves = new Set<StoreRecordId>();
             remove.forEach(id => {
-                if (!newRecords.has(id)) {
+                if (!eff(id)) {
                     missingRemoves++;
                     this.store.logDebug(`Attempted to remove non-existent record: ${id}`);
                     return;
                 }
                 allRemoves.add(id);
-                this.gatherDescendantIds(id, allRemoves);
+                if (isTree) this.gatherDescendantIds(id, allRemoves);
             });
-            allRemoves.forEach(it => newRecords.delete(it));
+            allRemoves.forEach(id => {
+                const rec = eff(id);
+                if (!rec) return;
+                count--;
+                if (rec.parentId == null) rootCount--;
+                removeFromPatch(id, newPatch, base);
+            });
         }
 
         // 1) Updates
         if (update) {
             update.forEach(rec => {
                 const {id} = rec,
-                    existing = newRecords.get(id);
+                    existing = eff(id);
                 if (!existing) {
                     missingUpdates++;
                     this.store.logDebug(`Attempted to update non-existent record: ${id}`);
                     return;
                 }
-                newRecords.set(id, rec);
+                newPatch.set(id, rec);
                 rec.finalize();
+                if (existing.parentId == null) rootCount--;
+                if (rec.parentId == null) rootCount++;
             });
         }
 
@@ -318,9 +487,11 @@ export class RecordSet {
         if (add) {
             add.forEach(rec => {
                 const {id} = rec;
-                throwIf(newRecords.has(id), `Attempted to insert duplicate record: ${id}`);
-                newRecords.set(id, rec);
+                throwIf(eff(id), `Attempted to insert duplicate record: ${id}`);
+                newPatch.set(id, rec);
                 rec.finalize();
+                count++;
+                if (rec.parentId == null) rootCount++;
             });
         }
 
@@ -329,12 +500,76 @@ export class RecordSet {
         if (missingUpdates > 0)
             logWarn(`Failed to update ${missingUpdates} records not found by id`, this);
 
-        const ret = new RecordSet(this.store, newRecords);
+        const ret = this.create(base, newPatch, count, rootCount);
         ret.derivation = {
-            type: 'full',
+            type: ret.base === base ? 'patched' : 'flattened',
             update: (update?.length ?? 0) - missingUpdates,
             add: add?.length ?? 0,
-            remove: this.count + (add?.length ?? 0) - newRecords.size
+            remove: this.count + (add?.length ?? 0) - count
+        };
+        return ret;
+    }
+
+    // Incremental arm of withFilter - null when not applicable, directing the caller to the
+    // full pass. See withFilter docs for the applicability conditions.
+    private withFilterIncremental(filter: Filter, prevFiltered: RecordSet): RecordSet {
+        if (
+            !prevFiltered ||
+            !this.isSameFilter(filter, prevFiltered._filter) ||
+            this.count !== this.rootCount ||
+            prevFiltered.count !== prevFiltered.rootCount
+        ) {
+            return null;
+        }
+        const delta = this.deltaFrom(prevFiltered._filterSource);
+        if (!delta) return null;
+
+        const {store} = this,
+            test = filter.getTestFn(store),
+            fBase = prevFiltered.base,
+            newPatch: PatchMap = prevFiltered.patch ? new Map(prevFiltered.patch) : new Map();
+        let count = prevFiltered.count,
+            added = 0,
+            removed = 0,
+            updated = 0;
+
+        delta.remove.forEach(rec => {
+            const {id} = rec;
+            if (prevFiltered.getById(id)) {
+                removeFromPatch(id, newPatch, fBase);
+                count--;
+                removed++;
+            }
+        });
+        delta.update.forEach(rec => {
+            const {id} = rec,
+                present = !!prevFiltered.getById(id);
+            if (test(rec)) {
+                newPatch.set(id, rec);
+                if (!present) count++;
+                present ? updated++ : added++;
+            } else if (present) {
+                removeFromPatch(id, newPatch, fBase);
+                count--;
+                removed++;
+            }
+        });
+        delta.add.forEach(rec => {
+            if (test(rec)) {
+                newPatch.set(rec.id, rec);
+                count++;
+                added++;
+            }
+        });
+
+        if (!(added || removed || updated)) return prevFiltered;
+
+        const ret = this.create(fBase, newPatch, count, count);
+        ret.derivation = {
+            type: ret.base === fBase ? 'patched' : 'flattened',
+            update: updated,
+            add: added,
+            remove: removed
         };
         return ret;
     }
@@ -342,6 +577,33 @@ export class RecordSet {
     //------------------------
     // Implementation
     //------------------------
+    /**
+     * Construct over a base + patch,flattening into a fresh base when the patch has grown past cap.
+     * Note `base` is not always this instance's own - incremental filtering derives over the
+     * previous projection's base.
+     */
+    private create(
+        base: StoreRecordMap,
+        patch: PatchMap,
+        count: number,
+        rootCount: number
+    ): RecordSet {
+        const {store} = this;
+        return patch.size > this.patchRatio() * base.size
+            ? new RecordSet(store, applyPatch(base, patch), null, count, rootCount)
+            : new RecordSet(store, base, patch, count, rootCount);
+    }
+
+    private isSameFilter(f1: Filter, f2: Filter): boolean {
+        return f1 === f2 || f1?.equals(f2);
+    }
+
+    /** Max patch size as a fraction of base, clamped to [0, 0.5] - see `experimental.maxPatchRatio`. */
+    private patchRatio(): number {
+        const ratio = this.store.experimental.maxPatchRatio;
+        return ratio == null ? DEFAULT_MAX_PATCH_RATIO : clamp(ratio, 0, MAX_PATCH_RATIO);
+    }
+
     private areRecordsEqual(r1: StoreRecord, r2: StoreRecord): boolean {
         if (r1 === r2) return true;
 
@@ -375,9 +637,9 @@ export class RecordSet {
         );
     }
 
-    private computeChildrenMap(recordMap: StoreRecordMap): ChildRecordMap {
+    private computeChildrenMap(): ChildRecordMap {
         const ret = new Map();
-        recordMap.forEach(r => {
+        this.forEachRecord(r => {
             const {parent} = r;
             if (parent) {
                 const children = ret.get(parent.id);
@@ -411,7 +673,23 @@ export class RecordSet {
     }
 }
 
-/** How a RecordSet instance was derived - counts only, stamped by Store. @internal */
+//------------------------
+// Patch-layer helpers - pure functions over explicit (base, patch) pairs, which do not always
+// belong to the operating instance (incremental filtering works over the previous projection's).
+//------------------------
+/** Record a removal in a patch: tombstone base entries, drop patch-only adds outright. */
+function removeFromPatch(id: StoreRecordId, patch: PatchMap, base: StoreRecordMap) {
+    base.has(id) ? patch.set(id, TOMBSTONE) : patch.delete(id);
+}
+
+/** Flatten a patch into a fresh map - in-place updates and appended adds preserve order. */
+function applyPatch(base: StoreRecordMap, patch: PatchMap): StoreRecordMap {
+    const ret = new Map(base);
+    patch.forEach((v, id) => (v === TOMBSTONE ? ret.delete(id) : ret.set(id, v)));
+    return ret;
+}
+
+/** How a RecordSet instance was derived - counts only, consumed by diagnostics. @internal */
 export interface RecordSetDerivation {
     type: 'patched' | 'flattened' | 'full' | 'unchanged';
     update: number;
